api/requirements.txt
Flask==3.0.0
PyYAML==6.0.1
requests==2.31.0
redis==5.0.1
mysql-connector-python==8.2.0
